backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
exceptiongroup==1.3.0
fastapi==0.118.0
h11==0.16.0
idna==3.10
pydantic==2.11.10
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0

Pillow==10.2.0
qrcode==7.4.2

sqlalchemy
pymysql
email-validator
passlib[bcrypt]
bcrypt<4.0.0
python-jose[cryptography]
requests
python-multipart

python-dotenv
requests

reportlab

qrcode[pil]
PyJWT